Work on coupling, mainly the connection interface.
Right now, I have it at the point where, on a remote node, you type in a name, desc, and URL, and hit enter. It shows a list of remote global channels. Check off a few and click the button. Then the socket server, while constantly reporting back to the client, probes the remote node for its UUID (this is ugly), creates an entry on the remote node for the local one (this is ugly), saves the local entry for the remote node, fetches the remote channel list, loops through it to import channels (in case of many channels they are each reported in realtime; progressbar later on if necesary), saves the records, and returns node info/list of imported channels to the client.

TODO: connect to remote node via sockets (to actually couple)