Node-powered RabbitMQ filesystem queue (w/ tests)
Requests sent to the queue 'rpc.node' will be processed by Node
on the filesystem. Node does not do auth, so care should be taken
to make sure the user can't directly access this RPC interface
(they shouldn't be connected directly to RabbitMQ anyway).

The test suite requires that channels/1 and users/1 exist, and
also depends on a file existing at users/1/clock.pdf (it was the
first file I saw, honest...). It doesn't error yet but just make
sure that the stuff it makes exist :)