Before I'm done for the day, here's fs.info
It takes an array of 'paths' and returns a hashmap
relating filenames provided to info about them. The one
key that will always be there is "type", with possible
values 'file', 'folder', and 'missing'. (The entire
request will fail under the Ruby interface is one path
is not allowed.)

Files and folders both have uploaded and added
timestamps, in milliseconds since the Unix epoche, as
well as the id of the original uploader. Files also have
mime type and file size in bytes.